+++ Sources/SwiftyHaru/DrawingContext.swift
@@ -110,7 +110,8 @@ public final class DrawingContext {
             return HPDF_Page_GetMiterLimit(_page)
         }
         set {
-            guard newValue >= 1 else { return }
+            precondition(newValue >= 1, "The minimum value of miterLimit is 1.0.")
+
             HPDF_Page_SetMiterLimit(_page, newValue)
         }
     }
@@ -118,7 +119,7 @@ public final class DrawingContext {
     /// The number of the page's graphics state stack.
     ///
     /// This number is increased whenever you call `withNewGState(_:)` or
-    /// `clip(to:evenOddRule:_:)` and decreased as soon as such a function returns.
+    /// `clip(to:rule:_:)` and decreased as soon as such a function returns.
     ///
     /// The maximum value of this property is `DrawingContext.maxGraphicsStateDepth`.
     public var graphicsStateDepth: Int {
@@ -328,14 +329,14 @@ public final class DrawingContext {
     ///              returns it is black again.
     ///
     /// - parameter path:                   The path that constraints the clipping area. Must be closed.
-    /// - parameter evenOddRule:            If `true`, uses even-odd rule for specifying a clipping area.
-    ///                                     Otherwise uses nonzero winding number rule.
+    /// - parameter rule:                   The rule for determining which areas to treat as the interior
+    ///                                     of the path. Default value is `.winding`.
     /// - parameter drawInsideClippingArea: All that is drawn inside this closure is clipped to the
     ///                                     provided `path`.
     /// - Throws:         `PDFError.exceedGStateLimit` if `graphicsStateDepth` is greater than
     ///                   `DrawingContext.maxGraphicsStateDepth`;
     ///                   rethrows errors thrown by the`drawInsideClippingArea` block.
-    public func clip(to path: Path, evenOddRule: Bool = false,
+    public func clip(to path: Path, rule: Path.FillRule = .winding,
                      _ drawInsideClippingArea: () throws -> Void) throws {
         
         try withNewGState {
@@ -344,9 +345,10 @@ public final class DrawingContext {
             
             HPDF_Page_ClosePath(_page)
             
-            if evenOddRule {
+            switch rule {
+            case .evenOdd:
                 HPDF_Page_Eoclip(_page)
-            } else {
+            case .winding:
                 HPDF_Page_Clip(_page)
             }
             
@@ -359,11 +361,10 @@ public final class DrawingContext {
     /// Fills the `path`.
     ///
     /// - parameter path:        The path to fill.
-    /// - parameter evenOddRule: If specified `true`, fills the path using the even-odd rule.
-    ///                          Otherwise fills it using the nonzero winding number rule.
-    ///                          Default value is `false`.
+    /// - parameter rule:        The rule for determining which areas to treat as the interior
+    ///                          of the path. Default value is `.winding`.
     /// - parameter stroke:      If specified `true`, also paints the path itself. Default value is `false`.
-    public func fill(_ path: Path, evenOddRule: Bool = false, stroke: Bool = false) {
+    public func fill(_ path: Path, rule: Path.FillRule = .winding, stroke: Bool = false) {
         
         assert(Int32(HPDF_Page_GetGMode(_page)) == HPDF_GMODE_PAGE_DESCRIPTION)
         
@@ -371,14 +372,14 @@ public final class DrawingContext {
         
         assert(Int32(HPDF_Page_GetGMode(_page)) == HPDF_GMODE_PATH_OBJECT)
         
-        switch (evenOddRule, stroke) {
-        case (true, true):
+        switch (rule, stroke) {
+        case (.evenOdd, true):
             HPDF_Page_EofillStroke(_page)
-        case (true, false):
+        case (.evenOdd, false):
             HPDF_Page_Eofill(_page)
-        case (false, true):
+        case (.winding, true):
             HPDF_Page_FillStroke(_page)
-        case (false, false):
+        case (.winding, false):
             HPDF_Page_Fill(_page)
         }
         
@@ -403,14 +404,31 @@ public final class DrawingContext {
     
     // MARK: - Text state
     
-    /// Tha current font of the context.
+    /// The current font of the context.
+    ///
+    /// You can only set fonts that has previously been loaded in the document using
+    /// `loadTrueTypeFont(from:embeddingGlyphData:)` or
+    /// `loadTrueTypeFontFromCollection(from:index:embeddingGlyphData:)` methods, or
+    /// ones predefined in the `Font` struct.
     public var font: Font {
         get {
             let fontHandle = HPDF_Page_GetCurrentFont(_page)
             return Font(name: String(cString: HPDF_Font_GetFontName(fontHandle)))
         }
         set {
-            let font = HPDF_GetFont(_documentHandle, newValue.name, encoding.name)
+            guard let font = HPDF_GetFont(_documentHandle, newValue.name, encoding.name) else {
+                
+                switch _document?._error {
+                case .some(PDFError.invalidFontName):
+                    preconditionFailure("Font \(newValue) must be loaded in the document using " +
+                        "loadTrueTypeFont(from:embeddingGlyphData:) or " +
+                        "loadTrueTypeFontFromCollection(from:index:embeddingGlyphData:) methods.")
+                case .some(let error):
+                    preconditionFailure(error.description)
+                default:
+                    return
+                }
+            }
             
             HPDF_Page_SetFontAndSize(_page, font, fontSize)
         }
@@ -422,13 +440,15 @@ public final class DrawingContext {
     }
     
     /// The size of the current font of the context. Valid values are between 0 and `maximumFontSize`.
-    /// Setting an invalid value makes mo change. Default value is 11.
+    /// Default value is 11.
     public var fontSize: Float {
         get {
             return HPDF_Page_GetCurrentFontSize(_page)
         }
         set {
-            guard newValue > 0 && newValue < maximumFontSize else { return }
+            
+            precondition(newValue > 0 && newValue < maximumFontSize, "Valid values for fontSize are " +
+                "between 0 and \(maximumFontSize).")
             
             let font = HPDF_GetFont(_documentHandle, self.font.name, encoding.name)