+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Please note that this version requires Swift 3.1.
 
+**Breaking changes:**
+
+- Changed the signature of the `DrawingContext.fill(_:evenOddRule:stroke:)` and `DrawingContext.clip(to:evenOddRule:_:)` methods. Instead of `Bool` value for fill rule they now require the `Path.FillRule` enum value.
+
 **Closed issues:**
 
 - Implement high-level grid drawing [\#6](https://github.com/WeirdMath/SwiftyHaru/issues/6)
@@ -24,6 +28,10 @@ Please note that this version requires Swift 3.1.
 - Closure-based syntax for saving and restoring the graphics state
 - Showing a text in a provided rectangle
 - **High-level customizable interface for drawing grids!**
+
+**Fixes:**
+
+- Setting `miterLimit`, `font` and `fontSize` of `DrawingContext` to an invalid value now causes precondition failure.
 
 ## [0.1.0](https://github.com/WeirdMath/SwiftyHaru/tree/0.1.0) (2016-11-02)
 **Closed issues:**